Move autoconfigure api to compileOnly dependency for all SPI implementation provider modules
