+++ exporters/logging-otlp/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 
 description = "OpenTelemetry Protocol JSON Logging Exporters"
 otelJava.moduleName.set("io.opentelemetry.exporter.logging.otlp")
+otelJava.osgiOptionalPackages.set(listOf("io.opentelemetry.api.incubator", "io.opentelemetry.sdk.autoconfigure.spi"))
 otelJava.osgiServiceLoaderProvides.set(listOf(
   "io.opentelemetry.sdk.autoconfigure.spi.traces.ConfigurableSpanExporterProvider",
   "io.opentelemetry.sdk.autoconfigure.spi.metrics.ConfigurableMetricExporterProvider",
@@ -21,12 +22,13 @@ dependencies {
 
   implementation(project(":exporters:otlp:common"))
   compileOnly(project(":api:incubator"))
-  implementation(project(":sdk-extensions:autoconfigure-spi"))
+  compileOnly(project(":sdk-extensions:autoconfigure-spi"))
 
   implementation("com.fasterxml.jackson.core:jackson-core")
 
   testImplementation(project(":api:incubator"))
   testImplementation(project(":sdk:testing"))
+  testImplementation(project(":sdk-extensions:autoconfigure-spi"))
 
   testImplementation("com.google.guava:guava")
   testImplementation("org.skyscreamer:jsonassert")
