+++ exporters/logging/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 
 description = "OpenTelemetry - Logging Exporter"
 otelJava.moduleName.set("io.opentelemetry.exporter.logging")
+otelJava.osgiOptionalPackages.set(listOf("io.opentelemetry.api.incubator", "io.opentelemetry.sdk.autoconfigure.spi"))
 otelJava.osgiServiceLoaderProvides.set(listOf(
   "io.opentelemetry.sdk.autoconfigure.spi.traces.ConfigurableSpanExporterProvider",
   "io.opentelemetry.sdk.autoconfigure.spi.metrics.ConfigurableMetricExporterProvider",
@@ -17,7 +18,7 @@ otelJava.osgiServiceLoaderProvides.set(listOf(
 dependencies {
   api(project(":sdk:all"))
 
-  implementation(project(":sdk-extensions:autoconfigure-spi"))
+  compileOnly(project(":sdk-extensions:autoconfigure-spi"))
   compileOnly(project(":api:incubator"))
 
   testImplementation(project(":sdk:testing"))
