+++ exporters/otlp/all/build.gradle.kts
@@ -9,7 +9,7 @@ apply<io.opentelemetry.gradle.OtelVersionClassPlugin>()
 
 description = "OpenTelemetry Protocol (OTLP) Exporters"
 otelJava.moduleName.set("io.opentelemetry.exporter.otlp")
-otelJava.osgiOptionalPackages.set(listOf("io.opentelemetry.api.incubator.config"))
+otelJava.osgiOptionalPackages.set(listOf("io.opentelemetry.api.incubator.config", "io.opentelemetry.sdk.autoconfigure.spi"))
 // io.grpc and org.jspecify.annotations are not OSGi bundles; must use unversioned optional.
 otelJava.osgiUnversionedOptionalPackages.set(listOf("io.grpc", "org.jspecify.annotations"))
 otelJava.osgiServiceLoaderProvides.set(listOf(
@@ -27,7 +27,7 @@ dependencies {
 
   implementation(project(":exporters:otlp:common"))
   implementation(project(":exporters:sender:okhttp"))
-  implementation(project(":sdk-extensions:autoconfigure-spi"))
+  compileOnly(project(":sdk-extensions:autoconfigure-spi"))
 
   compileOnly(project(":api:incubator"))
 
@@ -36,6 +36,7 @@ dependencies {
   compileOnly("io.grpc:grpc-stub")
 
   testImplementation(project(":exporters:otlp:testing-internal"))
+  testImplementation(project(":sdk-extensions:autoconfigure-spi"))
   testImplementation("com.linecorp.armeria:armeria-junit5")
   testImplementation("com.linecorp.armeria:armeria-grpc-protocol")
   testImplementation("io.grpc:grpc-stub")
