+++ exporters/zipkin/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 
 description = "OpenTelemetry - Zipkin Exporter"
 otelJava.moduleName.set("io.opentelemetry.exporter.zipkin")
+otelJava.osgiOptionalPackages.set(listOf("io.opentelemetry.sdk.autoconfigure.spi"))
 otelJava.osgiServiceLoaderProvides.set(listOf(
   "io.opentelemetry.sdk.autoconfigure.spi.traces.ConfigurableSpanExporterProvider",
 ))
@@ -17,12 +18,13 @@ dependencies {
   api("io.zipkin.reporter2:zipkin-reporter")
 
   implementation(project(":exporters:common"))
-  implementation(project(":sdk-extensions:autoconfigure-spi"))
+  compileOnly(project(":sdk-extensions:autoconfigure-spi"))
   compileOnly(project(":api:incubator"))
 
   implementation("io.zipkin.reporter2:zipkin-sender-okhttp3")
 
   testImplementation(project(":sdk:testing"))
+  testImplementation(project(":sdk-extensions:autoconfigure-spi"))
 
   testImplementation("com.linecorp.armeria:armeria")
   testImplementation("org.testcontainers:testcontainers-junit-jupiter")
