+++ extensions/trace-propagators/build.gradle.kts
@@ -8,7 +8,7 @@ plugins {
 
 description = "OpenTelemetry Extension : Trace Propagators"
 otelJava.moduleName.set("io.opentelemetry.extension.trace.propagation")
-otelJava.osgiOptionalPackages.set(listOf("io.opentelemetry.api.incubator"))
+otelJava.osgiOptionalPackages.set(listOf("io.opentelemetry.api.incubator", "io.opentelemetry.sdk.autoconfigure.spi"))
 otelJava.osgiServiceLoaderProvides.set(listOf(
   "io.opentelemetry.sdk.autoconfigure.spi.ConfigurablePropagatorProvider",
   "io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider",
